Moving map helper files into a separate folder
