+++ src/components/map_helpers/map_styles.js
